refactor shared utilities for triangular matrices
