+++ tpl-implementations/include/experimental/__p1673_bits/kokkos-kernels/blas2_matrix_rank_2_update.hpp
@@ -47,6 +47,7 @@
 #include <complex>
 #include "signal_kokkos_impl_called.hpp"
 #include "static_extent_match.hpp"
+#include "triangle.hpp"
 
 namespace KokkosKernelsSTD {
 
@@ -111,41 +112,6 @@ private:
   size_t ext1;
 };
 
-// Note: phrase it simply and the same as in specification ("has unique layout")
-template <typename Layout,
-          std::experimental::extents<>::size_type numRows,
-          std::experimental::extents<>::size_type numCols>
-
-inline constexpr bool is_unique_layout_v = Layout::template mapping<
-    std::experimental::extents<numRows, numCols> >::is_always_unique();
-
-template <typename Layout>
-struct is_layout_blas_packed: public std::false_type {};
-
-template <typename Triangle, typename StorageOrder>
-struct is_layout_blas_packed<
-  std::experimental::linalg::layout_blas_packed<Triangle, StorageOrder>>:
-    public std::true_type {};
-
-template <typename Layout>
-inline constexpr bool is_layout_blas_packed_v = is_layout_blas_packed<Layout>::value;
-
-// Note: will only signal failure for layout_blas_packed with different triangle
-template <typename Layout, typename Triangle>
-struct triangle_layout_match: public std::true_type {};
-
-template <typename StorageOrder, typename Triangle1, typename Triangle2>
-struct triangle_layout_match<
-  std::experimental::linalg::layout_blas_packed<Triangle1, StorageOrder>,
-  Triangle2>
-{
-  static constexpr bool value = std::is_same_v<Triangle1, Triangle2>;
-};
-
-template <typename Layout, typename Triangle>
-inline constexpr bool triangle_layout_match_v = triangle_layout_match<Layout, Triangle>::value;
-
-
 } // namespace mtxr2update_impl
 
 // Rank-2 update of a symmetric matrix
@@ -163,8 +129,8 @@ template<class ExecSpace,
          std::experimental::extents<>::size_type numCols_A,
          class Layout_A,
          class Triangle>
-  requires (mtxr2update_impl::is_unique_layout_v<Layout_A, numRows_A, numCols_A>
-            or mtxr2update_impl::is_layout_blas_packed_v<Layout_A>)
+  requires (Impl::is_unique_layout_v<Layout_A, numRows_A, numCols_A>
+            or Impl::is_layout_blas_packed_v<Layout_A>)
 void symmetric_matrix_rank_2_update(kokkos_exec<ExecSpace> &&exec,
   std::experimental::mdspan<ElementType_x, std::experimental::extents<ext_x>, Layout_x,
     std::experimental::default_accessor<ElementType_x>> x,
@@ -178,7 +144,7 @@ void symmetric_matrix_rank_2_update(kokkos_exec<ExecSpace> &&exec,
   static_assert(A.rank() == 2);
   static_assert(x.rank() == 1);
   static_assert(y.rank() == 1);
-  static_assert(mtxr2update_impl::triangle_layout_match_v<Layout_A, Triangle>);
+  static_assert(Impl::triangle_layout_match_v<Layout_A, Triangle>);
 
   // P1673 mandates
   static_assert(Impl::static_extent_match(A.static_extent(0), A.static_extent(1)));
@@ -224,8 +190,8 @@ template<class ExecSpace,
          std::experimental::extents<>::size_type numCols_A,
          class Layout_A,
          class Triangle>
-  requires (mtxr2update_impl::is_unique_layout_v<Layout_A, numRows_A, numCols_A>
-            or mtxr2update_impl::is_layout_blas_packed_v<Layout_A>)
+  requires (Impl::is_unique_layout_v<Layout_A, numRows_A, numCols_A>
+            or Impl::is_layout_blas_packed_v<Layout_A>)
 void hermitian_matrix_rank_2_update(kokkos_exec<ExecSpace> &&exec,
   std::experimental::mdspan<ElementType_x, std::experimental::extents<ext_x>, Layout_x,
     std::experimental::default_accessor<ElementType_x>> x,
@@ -239,7 +205,7 @@ void hermitian_matrix_rank_2_update(kokkos_exec<ExecSpace> &&exec,
   static_assert(A.rank() == 2);
   static_assert(x.rank() == 1);
   static_assert(y.rank() == 1);
-  static_assert(mtxr2update_impl::triangle_layout_match_v<Layout_A, Triangle>);
+  static_assert(Impl::triangle_layout_match_v<Layout_A, Triangle>);
 
   // P1673 mandates
   static_assert(Impl::static_extent_match(A.static_extent(0), A.static_extent(1)));
